requirements.txt
argcomplete==3.1.1
asgiref==3.7.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
boto3==1.28.20
botocore==1.31.20
certifi==2023.7.22
cfn-flip==1.3.0
charset-normalizer==3.2.0
click==8.1.6
dateutils==0.6.12
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==2.0.0
Django==4.2.3
django-admin-material-dashboard==1.0.10
django-admin-rangefilter==0.10.0
django-constance==2.9.1
django-filter==23.2
django-import-export==3.2.0
django-jet-reboot==1.3.3
django-picklefield==3.1
djangorestframework==3.14.0
durationpy==0.5
et-xmlfile==1.1.0
gunicorn==21.2.0
hjson==3.1.0
idna==3.4
importlib-metadata==6.8.0
jmespath==1.0.1
kappa==0.6.0
Markdown==3.4.3
MarkupPy==1.14
MarkupSafe==2.1.3
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
placebo==0.9.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-gettext==5.0
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.65.0
troposphere==4.3.2
typing_extensions==4.7.1
urllib3==1.26.16
Werkzeug==2.3.6
whitenoise==6.5.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.16.2
